增加通知提醒

+++ app/src/main/java/com/ql/recovery/yay/ui/dialog/MatchAudioDialog.kt
@@ -71,7 +71,7 @@ class MatchAudioDialog(
         return mUser
     }
 
-    fun setTime(time: Long) {
+    fun setTime(time: Int) {
         binding.tvTimer.text = time.toString()
     }
